feature/Movies2.0_001_movies Fixed tests for data layer
